refactor: modularize PanicHandler
